Moved main to app package
